fix: Fix Oauth, username in navbar and fetchs to localhost
I have fix the Oauth for google and 42 api, the problem was the call to the user-management microservice
which returned the id of the users instead of the real user.

Other change is the username in the navbar, that now is changed in all pages without reload, and it
also dissapears when the users logout.

The logout button now is in the navbar.

And now all the fetchs in login uses the host instead of localhost.